Add framebuffer module and move related functionality to it
